requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pyodbc==5.1.0
azure-identity==1.17.1
passlib[bcrypt]==1.7.4
pydantic==2.9.0
python-dotenv==1.0.1
email-validator
